refactor: move fn
